refactor: gas optimization

+++ contracts/common/lib/Unit.sol
@@ -2,21 +2,21 @@
 pragma solidity ^0.8;
 
 //------ units ------//
-uint constant LEVERAGE_UNIT = 10 ** 5;
-uint constant PROTOCOL_FEE_UNIT = 10 ** 3;
-uint constant SLIPPAGE_UNIT = 10 ** 3;
-uint constant APR_UNIT = 10 ** 2;
-uint constant REWARD_ERROR_UNIT = 10 ** 3;
-uint constant LIQUIDITY_ALLOWABLE_RATIO_UNIT = 10 ** 2;
-uint constant REWARD_PORTION_TOTAL_UNIT = 10 ** 3;
-uint constant SWAP_FEE_UNIT = 10 ** 3;
-uint constant LEND_RATIO_UNIT = 10 ** 3;
-uint constant SWAP_RATIO_UNIT = 10 ** 3;
+uint32 constant LEVERAGE_UNIT = 10 ** 5;
+uint32 constant PROTOCOL_FEE_UNIT = 10 ** 3;
+uint32 constant SLIPPAGE_UNIT = 10 ** 3;
+uint32 constant APR_UNIT = 10 ** 2;
+uint32 constant REWARD_ERROR_UNIT = 10 ** 3;
+uint32 constant LIQUIDITY_ALLOWABLE_RATIO_UNIT = 10 ** 2;
+uint32 constant REWARD_PORTION_TOTAL_UNIT = 10 ** 3;
+uint32 constant SWAP_FEE_UNIT = 10 ** 3;
+uint32 constant LEND_RATIO_UNIT = 10 ** 3;
+uint32 constant SWAP_RATIO_UNIT = 10 ** 3;
 
 
 //------ token Unit ------//
-uint constant INVI_UNIT = 10**18;
-uint constant ILP_UNIT = 10**18;
-uint constant INKLP_UNIT = 10 ** 18;
+uint128 constant INVI_UNIT = 10**18;
+uint128 constant ILP_UNIT = 10**18;
+uint128 constant INKLP_UNIT = 10 ** 18;
